+++ src/main/java/iris/task/Task.java
@@ -1,4 +1,4 @@
-package duke.task;
+package iris.task;
 
 /**
  * Represents a Task
