+++ text-ui-test/runtest.sh
@@ -28,7 +28,7 @@ then
 fi
 
 # run the program, feed commands from input.txt file and redirect the output to the ACTUAL.TXT
-java -classpath ../bin Duke < input.txt > ACTUAL.TXT
+java -classpath ../bin iris/Iris < input.txt > ACTUAL.TXT
 
 # convert to UNIX format
 #cp EXPECTED.TXT EXPECTED-UNIX.TXT
