+++ text-ui-test/sources.txt
@@ -1,6 +1,6 @@
-../src/main/java/duke/Duke.java
-../src/main/java/duke/IrisException.java
-../src/main/java/duke/Parser.java
-../src/main/java/duke/Storage.java
-../src/main/java/duke/TaskList.java
-../src/main/java/duke/Ui.java
+../src/main/java/iris/Iris.java
+../src/main/java/iris/IrisException.java
+../src/main/java/iris/Parser.java
+../src/main/java/iris/Storage.java
+../src/main/java/iris/TaskList.java
+../src/main/java/iris/Ui.java
